Upgrade lru cache
