+++ engine/access/rpc/connection/connection_test.go
@@ -4,31 +4,25 @@ import (
 	"context"
 	"fmt"
 	"net"
-	"strconv"
-	"strings"
 	"sync"
 	"testing"
 	"time"
 
-	"github.com/sony/gobreaker"
-
-	"go.uber.org/atomic"
-	"pgregory.net/rapid"
-
-	lru "github.com/hashicorp/golang-lru"
+	lru "github.com/hashicorp/golang-lru/v2"
+	"github.com/onflow/flow-go/module/metrics"
+	"github.com/onflow/flow-go/utils/unittest"
 	"github.com/onflow/flow/protobuf/go/flow/access"
 	"github.com/onflow/flow/protobuf/go/flow/execution"
+	"github.com/sony/gobreaker"
 	"github.com/stretchr/testify/assert"
 	testifymock "github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
+	"go.uber.org/atomic"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/connectivity"
 	"google.golang.org/grpc/status"
-
-	"github.com/onflow/flow-go/engine/access/mock"
-	"github.com/onflow/flow-go/module/metrics"
-	"github.com/onflow/flow-go/utils/unittest"
+	"pgregory.net/rapid"
 )
 
 func TestProxyAccessAPI(t *testing.T) {
@@ -70,6 +64,14 @@ func TestProxyAccessAPI(t *testing.T) {
 	resp, err := client.Ping(ctx, req)
 	assert.NoError(t, err)
 	assert.Equal(t, resp, expected)
+}
+
+func getCache(t *testing.T, cacheSize int) *lru.Cache[string, *CachedClient] {
+	cache, err := lru.NewWithEvict[string, *CachedClient](cacheSize, func(_ string, client *CachedClient) {
+		client.Close()
+	})
+	require.NoError(t, err)
+	return cache
 }
 
 func TestProxyExecutionAPI(t *testing.T) {
@@ -129,10 +131,7 @@ func TestProxyAccessAPIConnectionReuse(t *testing.T) {
 	connectionFactory.CollectionGRPCPort = cn.port
 	// set the connection pool cache size
 	cacheSize := 1
-	cache, _ := lru.NewWithEvict(cacheSize, func(_, evictedValue interface{}) {
-		evictedValue.(*CachedClient).Close()
-	})
-	connectionCache := NewCache(cache, cacheSize)
+	connectionCache := NewCache(getCache(t, cacheSize), cacheSize)
 
 	// set metrics reporting
 	connectionFactory.AccessMetrics = metrics.NewNoopCollector()
@@ -184,10 +183,7 @@ func TestProxyExecutionAPIConnectionReuse(t *testing.T) {
 	connectionFactory.ExecutionGRPCPort = en.port
 	// set the connection pool cache size
 	cacheSize := 5
-	cache, _ := lru.NewWithEvict(cacheSize, func(_, evictedValue interface{}) {
-		evictedValue.(*CachedClient).Close()
-	})
-	connectionCache := NewCache(cache, cacheSize)
+	connectionCache := NewCache(getCache(t, cacheSize), cacheSize)
 	// set metrics reporting
 	connectionFactory.AccessMetrics = metrics.NewNoopCollector()
 	connectionFactory.Manager = NewManager(
@@ -245,10 +241,7 @@ func TestExecutionNodeClientTimeout(t *testing.T) {
 	connectionFactory.ExecutionNodeGRPCTimeout = timeout
 	// set the connection pool cache size
 	cacheSize := 5
-	cache, _ := lru.NewWithEvict(cacheSize, func(_, evictedValue interface{}) {
-		evictedValue.(*CachedClient).Close()
-	})
-	connectionCache := NewCache(cache, cacheSize)
+	connectionCache := NewCache(getCache(t, cacheSize), cacheSize)
 	// set metrics reporting
 	connectionFactory.AccessMetrics = metrics.NewNoopCollector()
 	connectionFactory.Manager = NewManager(
@@ -294,10 +287,7 @@ func TestCollectionNodeClientTimeout(t *testing.T) {
 	connectionFactory.CollectionNodeGRPCTimeout = timeout
 	// set the connection pool cache size
 	cacheSize := 5
-	cache, _ := lru.NewWithEvict(cacheSize, func(_, evictedValue interface{}) {
-		evictedValue.(*CachedClient).Close()
-	})
-	connectionCache := NewCache(cache, cacheSize)
+	connectionCache := NewCache(getCache(t, cacheSize), cacheSize)
 	// set metrics reporting
 	connectionFactory.AccessMetrics = metrics.NewNoopCollector()
 	connectionFactory.Manager = NewManager(
@@ -343,10 +333,8 @@ func TestConnectionPoolFull(t *testing.T) {
 	connectionFactory.CollectionGRPCPort = cn1.port
 	// set the connection pool cache size
 	cacheSize := 2
-	cache, _ := lru.NewWithEvict(cacheSize, func(_, evictedValue interface{}) {
-		evictedValue.(*CachedClient).Close()
-	})
-	connectionCache := NewCache(cache, cacheSize)
+	connectionCache := NewCache(getCache(t, cacheSize), cacheSize)
+
 	// set metrics reporting
 	connectionFactory.AccessMetrics = metrics.NewNoopCollector()
 	connectionFactory.Manager = NewManager(
@@ -419,10 +407,8 @@ func TestConnectionPoolStale(t *testing.T) {
 	connectionFactory.CollectionGRPCPort = cn.port
 	// set the connection pool cache size
 	cacheSize := 5
-	cache, _ := lru.NewWithEvict(cacheSize, func(_, evictedValue interface{}) {
-		evictedValue.(*CachedClient).Close()
-	})
-	connectionCache := NewCache(cache, cacheSize)
+	connectionCache := NewCache(getCache(t, cacheSize), cacheSize)
+
 	// set metrics reporting
 	connectionFactory.AccessMetrics = metrics.NewNoopCollector()
 	connectionFactory.Manager = NewManager(
@@ -488,7 +474,7 @@ func TestExecutionNodeClientClosedGracefully(t *testing.T) {
 	}
 
 	// Add rapid test, to check graceful close on different number of requests
-	rapid.Check(t, func(t *rapid.T) {
+	rapid.Check(t, func(tt *rapid.T) {
 		en, closer := createExecNode()
 		defer closer()
 
@@ -508,10 +494,8 @@ func TestExecutionNodeClientClosedGracefully(t *testing.T) {
 		connectionFactory.ExecutionNodeGRPCTimeout = time.Second
 		// set the connection pool cache size
 		cacheSize := 1
-		cache, _ := lru.NewWithEvict(cacheSize, func(_, evictedValue interface{}) {
-			evictedValue.(*CachedClient).Close()
-		})
-		connectionCache := NewCache(cache, cacheSize)
+		connectionCache := NewCache(getCache(t, cacheSize), cacheSize)
+
 		// set metrics reporting
 		connectionFactory.AccessMetrics = metrics.NewNoopCollector()
 		connectionFactory.Manager = NewManager(
@@ -530,7 +514,7 @@ func TestExecutionNodeClientClosedGracefully(t *testing.T) {
 		ctx := context.Background()
 
 		// Generate random number of requests
-		nofRequests := rapid.IntRange(10, 100).Draw(t, "nofRequests").(int)
+		nofRequests := rapid.IntRange(10, 100).Draw(tt, "nofRequests").(int)
 		reqCompleted := atomic.NewUint64(0)
 
 		var waitGroup sync.WaitGroup
@@ -592,10 +576,10 @@ func TestExecutionEvictingCacheClients(t *testing.T) {
 	connectionFactory.CollectionNodeGRPCTimeout = 5 * time.Second
 	// Set the connection pool cache size
 	cacheSize := 1
-	cache, err := lru.New(cacheSize)
+	cache, err := lru.New[string, *CachedClient](cacheSize)
 	require.NoError(t, err)
 
-	connectionCache := NewCache(cache, cacheSize)
+	connectionCache := NewCache(getCache(t, cacheSize), cacheSize)
 	// set metrics reporting
 	connectionFactory.AccessMetrics = metrics.NewNoopCollector()
 	connectionFactory.Manager = NewManager(
@@ -614,9 +598,8 @@ func TestExecutionEvictingCacheClients(t *testing.T) {
 	ctx := context.Background()
 
 	// Retrieve the cached client from the cache
-	result, _ := cache.Get(clientAddress)
-	cachedClient := result.(*CachedClient)
-
+	cachedClient, ok := cache.Get(clientAddress)
+	require.True(t, ok)
 	// Schedule the invalidation of the access API client after a delay
 	time.AfterFunc(250*time.Millisecond, func() {
 		// Invalidate the access API client
@@ -671,7 +654,7 @@ func TestCircuitBreakerExecutionNode(t *testing.T) {
 
 	// Set the connection pool cache size.
 	cacheSize := 1
-	connectionCache, _ := lru.New(cacheSize)
+	connectionCache, _ := lru.New[string, *CachedClient](cacheSize)
 
 	connectionFactory.Manager = NewManager(
 		NewCache(connectionCache, cacheSize),
@@ -755,7 +738,7 @@ func TestCircuitBreakerCollectionNode(t *testing.T) {
 
 	// Set the connection pool cache size.
 	cacheSize := 1
-	connectionCache, _ := lru.New(cacheSize)
+	connectionCache, _ := lru.New[string, *CachedClient](cacheSize)
 
 	connectionFactory.Manager = NewManager(
 		NewCache(connectionCache, cacheSize),
@@ -811,80 +794,4 @@ func TestCircuitBreakerCollectionNode(t *testing.T) {
 	duration, err = callAndMeasurePingDuration()
 	assert.Greater(t, requestTimeout, duration)
 	assert.Equal(t, nil, err)
-}
-
-// node mocks a flow node that runs a GRPC server
-type node struct {
-	server   *grpc.Server
-	listener net.Listener
-	port     uint
-}
-
-func (n *node) setupNode(t *testing.T) {
-	n.server = grpc.NewServer()
-	listener, err := net.Listen("tcp4", unittest.DefaultAddress)
-	assert.NoError(t, err)
-	n.listener = listener
-	assert.Eventually(t, func() bool {
-		return !strings.HasSuffix(listener.Addr().String(), ":0")
-	}, time.Second*4, 10*time.Millisecond)
-
-	_, port, err := net.SplitHostPort(listener.Addr().String())
-	assert.NoError(t, err)
-	portAsUint, err := strconv.ParseUint(port, 10, 32)
-	assert.NoError(t, err)
-	n.port = uint(portAsUint)
-}
-
-func (n *node) start(t *testing.T) {
-	// using a wait group here to ensure the goroutine has started before returning. Otherwise,
-	// there's a race condition where the server is sometimes stopped before it has started
-	wg := sync.WaitGroup{}
-	wg.Add(1)
-	go func() {
-		wg.Done()
-		err := n.server.Serve(n.listener)
-		assert.NoError(t, err)
-	}()
-	unittest.RequireReturnsBefore(t, wg.Wait, 10*time.Millisecond, "could not start goroutine on time")
-}
-
-func (n *node) stop(t *testing.T) {
-	if n.server != nil {
-		n.server.Stop()
-	}
-}
-
-type executionNode struct {
-	node
-	handler *mock.ExecutionAPIServer
-}
-
-func (en *executionNode) start(t *testing.T) {
-	en.setupNode(t)
-	handler := new(mock.ExecutionAPIServer)
-	execution.RegisterExecutionAPIServer(en.server, handler)
-	en.handler = handler
-	en.node.start(t)
-}
-
-func (en *executionNode) stop(t *testing.T) {
-	en.node.stop(t)
-}
-
-type collectionNode struct {
-	node
-	handler *mock.AccessAPIServer
-}
-
-func (cn *collectionNode) start(t *testing.T) {
-	cn.setupNode(t)
-	handler := new(mock.AccessAPIServer)
-	access.RegisterAccessAPIServer(cn.server, handler)
-	cn.handler = handler
-	cn.node.start(t)
-}
-
-func (cn *collectionNode) stop(t *testing.T) {
-	cn.node.stop(t)
 }
