+++ engine/consensus/approvals/approvals_lru_cache.go
@@ -3,7 +3,7 @@ package approvals
 import (
 	"sync"
 
-	"github.com/hashicorp/golang-lru/simplelru"
+	"github.com/hashicorp/golang-lru/v2/simplelru"
 
 	"github.com/onflow/flow-go/model/flow"
 )
@@ -11,7 +11,7 @@ import (
 // LruCache is a wrapper over `simplelru.LRUCache` that provides needed api for processing result approvals
 // Extends functionality of `simplelru.LRUCache` by introducing additional index for quicker access.
 type LruCache struct {
-	lru  simplelru.LRUCache
+	lru  simplelru.LRUCache[flow.Identifier, *flow.ResultApproval]
 	lock sync.RWMutex
 	// secondary index by result id, since multiple approvals could
 	// reference same result
@@ -21,8 +21,7 @@ type LruCache struct {
 func NewApprovalsLRUCache(limit uint) *LruCache {
 	byResultID := make(map[flow.Identifier]map[flow.Identifier]struct{})
 	// callback has to be called while we are holding lock
-	lru, _ := simplelru.NewLRU(int(limit), func(key interface{}, value interface{}) {
-		approval := value.(*flow.ResultApproval)
+	lru, _ := simplelru.NewLRU(int(limit), func(key flow.Identifier, approval *flow.ResultApproval) {
 		delete(byResultID[approval.Body.ExecutionResultID], approval.Body.PartialID())
 		if len(byResultID[approval.Body.ExecutionResultID]) == 0 {
 			delete(byResultID, approval.Body.ExecutionResultID)
@@ -40,7 +39,7 @@ func (c *LruCache) Peek(approvalID flow.Identifier) *flow.ResultApproval {
 	// check if we have it in the cache
 	resource, cached := c.lru.Peek(approvalID)
 	if cached {
-		return resource.(*flow.ResultApproval)
+		return resource
 	}
 
 	return nil
@@ -52,7 +51,7 @@ func (c *LruCache) Get(approvalID flow.Identifier) *flow.ResultApproval {
 	// check if we have it in the cache
 	resource, cached := c.lru.Get(approvalID)
 	if cached {
-		return resource.(*flow.ResultApproval)
+		return resource
 	}
 
 	return nil
@@ -74,7 +73,7 @@ func (c *LruCache) TakeByResultID(resultID flow.Identifier) []*flow.ResultApprov
 			// no need to cleanup secondary index since it will be
 			// cleaned up in evict callback
 			_ = c.lru.Remove(approvalID)
-			approvals = append(approvals, resource.(*flow.ResultApproval))
+			approvals = append(approvals, resource)
 		}
 	}
